Setting up Basis of Assignment-01
Using the express-generator module to create the basic setup of my app. Creating a .gitignore file, so that the node_modules folder will not be tracked (as the modules within it will be listed as dependancies in the package.json file, and so will be installed when this repo is cloned down and installed on another machine).