Using Try/Catch to allow app to run locally, and on Azure
As per my previous commit, Azure seems to think that ./bin/secrets.json is in '/bin/bin/secrets', but locally (i.e. on my computer) this is /bin/secrets.json is the correct path. To accomodate for this, I am using a try/catch statement to ensure that the right path is always available, on both Azure and locally