Using express-session to store username
Currently unsure as to how much I can completely away from cookies, as setting the session 'username' value to a new value, does not seem to be reflected in the cookie's stored in the browser, so for passing information that I want to be able to alter/update on the client side, I may need to stick with cookies i.e. did they want their secrets sorted by title or date?