Successfully running local app using remote database on Azure
Login credentials have been removed from the app for the purposes of this commit, but I can now run my app locally, while utilising a MySQL database I have running on Azure. Now I just need to figure out the last piece in connecting the Azure app with the Azure database (currently, the app references the connection string of the database, but I am still unsure of how I reference that within the code of the app i.e. what hostname etc do I pass in to the connection setup. Will work on it