Tidying Up
Moving the embedded JavaScript from the secrets.jade view, into the public script.js, so that I can harness the capabilities of jQuery, as well as seperating out the functionality from the display of the information. Instead of using a variable on the server to store the current sortBy value of the cookie, instead passing the value into the SQL query directly (as the benefit of cookies is that I don't have to store these values, as they will be passed back and forth). Using the relevant database headings from the Secret table, as the values for the select options for sorting, as it greatly reduces the amount of additional code required to decipher which method to sort by on the server side. Beginning lining up some of the elements on the screen e.g. logout button, overall div structure etc. Removing unnecessary require() statements from the user.js and index.js routes, as they were left over from previous methods, and are no longer needed.