Validating login/create account forms before they are submitted
Checking that forms have data entered into their username / password inputs, and checking if the confirm password input (on the create account tab) matches the data in the password input (of the same tab). These client side checks will save time and reduce the need for unnecessary requests to be sent when the relevant data has not been entered.