Changing domain of userPassword in the database
Figured out why I couldn't create/login to a new account on the remote Azure database. The domain of the userPassword field in the User table was VARCHAR(45). Since I am encrypting all passwords etc, the resulting value will not fit with this field, so I have changed it to VARBINARY(400) (as I have previously done with the secretTitle and secretDescription fields in the Secret table. I can now access and interact with the remote database (albeit using the full connection string with the hostname, username, password and database name - which has been removed from this commit for security purposes.