Successfully sorting case insensitive
Successfully sorting case insensitive. There were a few hurdles to this as 1) the column I want to sort by in encrypted, 2) the domain of the column I want to sort by is binary and 3) I wanted to be able to pass the sorting method to be used directly from the req.cookie, instead of having to create a temporary variable. I am now using a ternary operator to set the temporary variable to be equal to the relevant sorting instructions, as the secretTitle requires alot of additional parametres to make it sort case insensitive, with capital letters first. Now working successfully.