Using Environment Variables to store encryption keys
Using environment variables to store encryption keys, so that they are no longer public on GitHub. Once I have tested this on Azure, I will restart and clear the database, before changing all of the values for these keys to new unique values. (I will have to restart the database because any encrypted data in it would be lost if I changed the keys while there was still rows in any of the tables.